separating app.js into new folder structure
