+++ src/config.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+  PORT: process.env.PORT || 4000,
+  NODE_ENV: process.env.NODE_ENV || 'development',
+}
