+++ src/server.js
@@ -1,0 +1,18 @@
+const app = require('./app');
+const { PORT } = require('./config');
+const { ApolloServer } = require('apollo-server-express');
+const { typeDefs } = require('./typeDefs/index')
+const { resolvers } = require('./resolvers/index')
+
+const server = new ApolloServer({ 
+  typeDefs, 
+  resolvers, 
+  introspection: true, 
+  playground: true 
+});
+
+server.applyMiddleware({ app });
+
+app.listen(PORT, () => {
+  console.log(`Server started at http://localhost:4000${server.graphqlPath}`);
+})
